+++ tda/orders/__init__.py
@@ -2,7 +2,7 @@ from enum import Enum
 
 from . import common
 from . import equities
-from . import complex
+from . import generic
 
 __WARNINGS = set()
